Move lerpGradient function from utils to lerp.dart
